+++ public_html/js/views/GameModules/gameObjects.js
@@ -1,0 +1,13 @@
+define(function (require) {
+
+    var objects = {
+        scene: null,
+        camera: null,
+        light: null,
+        renderer: null,
+        
+    };
+    
+    return objects;
+
+});
